+++ mbed_cloud_client_user_config.h
@@ -1,5 +1,5 @@
 // ----------------------------------------------------------------------------
-// Copyright 2016-2017 ARM Ltd.
+// Copyright 2016-2019 ARM Ltd.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -20,21 +20,20 @@
 #ifndef MBED_CLOUD_CLIENT_USER_CONFIG_H
 #define MBED_CLOUD_CLIENT_USER_CONFIG_H
 
-#define MBED_CLOUD_CLIENT_ENDPOINT_TYPE         "default"
-
-#define MBED_CLOUD_CLIENT_LIFETIME              3600
+#define MBED_CLOUD_CLIENT_ENDPOINT_TYPE             "default"
+#define MBED_CLOUD_CLIENT_LIFETIME                  3600
 
 #ifdef __FREERTOS__
-    #define SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE       512
+    #define SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE      512
     #define MBED_CLOUD_CLIENT_TRANSPORT_MODE_TCP
 #elif defined(__SXOS__)
-    #define SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE       512
+    #define SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE      512
     #define MBED_CLOUD_CLIENT_TRANSPORT_MODE_UDP_QUEUE
 #else
 #ifdef MBED_CONF_MBED_CLIENT_SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE
     #define SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE    MBED_CONF_MBED_CLIENT_SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE
 #else
-    #define SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE 1024
+    #define SN_COAP_MAX_BLOCKWISE_PAYLOAD_SIZE      512
 #endif
     #define MBED_CLOUD_CLIENT_TRANSPORT_MODE_TCP
 #endif
@@ -43,13 +42,13 @@
 #if defined(__linux__) || defined(TARGET_LIKE_MBED) || defined(__SXOS__)
     #define MBED_CLOUD_CLIENT_SUPPORT_UPDATE
 #endif
-/* set download buffer size in bytes (min. 1024 bytes) */
 
+/* set download buffer size in bytes (min. 1024 bytes) */
 /* Use larger buffers in Linux */
 #ifdef __linux__
-#define MBED_CLOUD_CLIENT_UPDATE_BUFFER          (2 * 1024 * 1024)
+#define MBED_CLOUD_CLIENT_UPDATE_BUFFER             (2 * 1024 * 1024)
 #else
-#define MBED_CLOUD_CLIENT_UPDATE_BUFFER          2048
+#define MBED_CLOUD_CLIENT_UPDATE_BUFFER             2048
 #endif
 
 /* Developer flags for Update feature */