+++ source/platform/Linux/mcc_common_button_and_led.c
@@ -1,18 +1,20 @@
-/*
- * Copyright (c) 2015-2018 ARM Limited. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- * Licensed under the Apache License, Version 2.0 (the License); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an AS IS BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// ----------------------------------------------------------------------------
+// Copyright 2015-2019 ARM Ltd.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------
 
 ///////////
 // INCLUDES
@@ -59,10 +61,13 @@ uint8_t mcc_platform_button_clicked(void)
 #if PLATFORM_ENABLE_BUTTON
 void *button_thread(void *arg)
 {
-    for(;;) {
-        getchar();
+    // EOF checking is required, otherwise this ends up as a busy loop
+    // and consumes a lot of CPU. EOF will be return if there is no tty
+    // or Ctrl+D is pressed.
+    while (getchar() != EOF) {
         button_pressed = 1;
     }
+    button_pressed = 0;
     return NULL;
 }
 #endif